+++ tests/integration/home.test.js
@@ -17,4 +17,14 @@ test(function (test) {
 
     test.itShould.containLinkTo('asimovjs.org');
   });
+
+  test.integration('/the-alias-for-home', function () {
+
+    test.itShould.loadPage();
+
+    test.it('should have the correct title', function () {
+
+      expect(this.browser.text('H1')).to.contain('asimov.js');
+    });
+  });
 });