+++ CHANGELOG.md
@@ -5,3 +5,4 @@
   - Fixed high CPU usage from watching directories [#165](https://github.com/adamrenklint/asimov.js/issues/165)
   - Better error message when trying to include a non-existing stylesheet [#166](https://github.com/adamrenklint/asimov.js/issues/166)
   - Handle both single and double quoted @import statements [#168](https://github.com/adamrenklint/asimov.js/issues/168)
+  - Simple redirects with page.attributes.alias [#45](https://github.com/adamrenklint/asimov.js/issues/45)